Added a class when a menu item is active
Changed the application helper so a class "active" is added whenever the menu item is active. (user is on the page)

And also changed init.rb so that the server reloads automatically when some file changes, the lines that do this were commented out. But it is a pain when developing. There is a check for environment, so it should only be the case in the development environment.